@@ -14,15 +14,15 @@
         }
     ],
     "require": {
-        "php": ">=7.4 <8.2",
-        "atk4/data": "dev-develop",
-        "atk4/ui": "dev-develop",
+        "php": "< 8.3 >=7.4",
+        "atk4/data": "~4.0.0",
+        "atk4/ui": "~4.0.0",
         "phpmailer/phpmailer": "^v6.5.3"
     },
     "require-release": {
-        "php": ">=7.4 <8.2",
-        "atk4/data": "~3.2.0",
-        "atk4/ui": "~3.2.0",
+        "php": ">=7.4 < 8.3",
+        "atk4/data": "~4.0.0",
+        "atk4/ui": "~4.0.0",
         "phpmailer/phpmailer": "^v6.5.3"
     },
     "require-dev": {
@@ -48,6 +48,10 @@
         }
     },
     "config": {
+        "allow-plugins": {
+            "ergebnis/composer-normalize": true,
+            "phpstan/extension-installer": true
+        },
         "sort-packages": true
     }
 }
