@@ -2,17 +2,15 @@
 
 $finder = PhpCsFixer\Finder::create()
     ->in([__DIR__])
-    ->exclude([
-        'vendor',
-    ]);
+    ->exclude(['vendor']);
 
-$config = new \PhpCsFixer\Config();
-$config->setRiskyAllowed(true)
+return (new PhpCsFixer\Config())
+    ->setRiskyAllowed(true)
     ->setRules([
         '@PhpCsFixer' => true,
-        '@PhpCsFixer:risky' =>true,
-        '@PHP71Migration:risky' => true,
-        '@PHP73Migration' => true,
+        '@PhpCsFixer:risky' => true,
+        '@PHP74Migration:risky' => true,
+        '@PHP74Migration' => true,
 
         // required by PSR-12
         'concat_space' => [
@@ -23,7 +21,7 @@ $config->setRiskyAllowed(true)
         'phpdoc_types' => [
             // keep enabled, but without "alias" group to not fix
             // "Callback" to "callback" in phpdoc
-            'groups' => ['simple', 'meta']
+            'groups' => ['simple', 'meta'],
         ],
         'phpdoc_types_order' => [
             'null_adjustment' => 'always_last',
@@ -34,10 +32,8 @@ $config->setRiskyAllowed(true)
             'equal' => false,
             'identical' => false,
         ],
+        'native_constant_invocation' => true,
         'native_function_invocation' => false,
-        'non_printable_character' => [
-            'use_escape_sequences_in_strings' => true,
-        ],
         'void_return' => false,
         'blank_line_before_statement' => [
             'statements' => ['break', 'continue', 'declare', 'return', 'throw', 'exit'],
@@ -48,22 +44,20 @@ $config->setRiskyAllowed(true)
         'no_superfluous_elseif' => false,
         'ordered_class_elements' => false,
         'php_unit_internal_class' => false,
-        'php_unit_test_case_static_method_calls' => [
-            'call_type' => 'this',
-        ],
         'php_unit_test_class_requires_covers' => false,
         'phpdoc_add_missing_param_annotation' => false,
         'return_assignment' => false,
         'comment_to_phpdoc' => false,
-        'list_syntax' => ['syntax' => 'short'],
         'general_phpdoc_annotation_remove' => [
             'annotations' => ['author', 'copyright', 'throws'],
         ],
         'nullable_type_declaration_for_default_null_value' => [
             'use_nullable_type_declaration' => false,
         ],
+
+        // fn => without curly brackets is less readable,
+        // also prevent bounding of unwanted variables for GC
+        'use_arrow_functions' => false,
     ])
     ->setFinder($finder)
-    ->setCacheFile(__DIR__ . '/.php_cs.cache');
-
-return $config;
+    ->setCacheFile(sys_get_temp_dir() . '/php-cs-fixer.' . md5(__DIR__) . '.cache');
