@@ -17,9 +17,10 @@ $sqliteFile = __DIR__ . '/db.sqlite';
 if (!file_exists($sqliteFile)) {
     new Persistence\Sql('sqlite:' . $sqliteFile);
 }
+
 unset($sqliteFile);
 
-/** @var Persistence\Sql $db */
+/** @var Persistence|Persistence\Sql $db */
 require_once __DIR__ . '/../init-db.php';
 
 echo 'GITHUB_JOB : ' . getenv('GITHUB_JOB') . "\n\n";
