@@ -20,18 +20,19 @@ $app = new App();
 
 // collect coverage for HTTP tests 2/2
 if (file_exists(__DIR__ . '/CoverageUtil.php') && !class_exists(\PHPUnit\Framework\TestCase::class, false)) {
-    $app->onHook(\Atk4\Ui\App::HOOK_BEFORE_EXIT, function () {
+    $app->onHook(\Atk4\Ui\App::HOOK_BEFORE_EXIT, static function () {
         \CoverageUtil::saveData();
     });
 }
 
 try {
-    /** @var Persistence\Sql $db */
+    /** @var Persistence|Persistence\Sql $db */
     require_once __DIR__ . '/init-db.php';
+
     $app->db = $db;
     unset($db);
-} catch (\Throwable $e) {
-    throw new \Atk4\Ui\Exception('Database error: ' . $e->getMessage());
+} catch (\Throwable $throwable) {
+    throw new \Atk4\Ui\Exception('Database error: ' . $throwable->getMessage());
 }
 
 $app->invokeInit();
