@@ -8,8 +8,8 @@ try {
     require_once file_exists(__DIR__ . '/db.php')
         ? __DIR__ . '/db.php'
         : __DIR__ . '/db.default.php';
-} catch (\PDOException $e) {
+} catch (\PDOException $pdoException) {
     // do not show $e unless you can secure DSN!
     throw (new \Atk4\Ui\Exception('This demo requires access to the database. See "demos/init-db.php"'))
-        ->addMoreInfo('PDO error', $e->getMessage());
+        ->addMoreInfo('PDO error', $pdoException->getMessage());
 }
