@@ -9,7 +9,10 @@ use Atk4\Data\Model;
 class MailTemplate extends Model
 {
     public $table = 'mail_template';
+
     public $caption = 'Mail Template';
+
+    public ?string $titleField = 'identifier';
 
     protected function init(): void
     {
@@ -36,46 +39,49 @@ class MailTemplate extends Model
 
         $this->containsMany('tokens', ['model' => [MailTemplateToken::class]]);
 
-        $this->onHook('beforeSave', function (self $m) {
+        $this->onHook(Model::HOOK_AFTER_SAVE, static function (self $m) {
             $m->refreshTokens();
         }, [], -200);
     }
 
     public function refreshTokens(): void
     {
+        $original_tokens = $this->get('tokens') ?? [];
+        $tokens = [];
+        foreach ($original_tokens as $token) {
+            $tokens[$token['token']] = [
+                'token' => $token['token'],
+                'description' => $token['description'],
+            ];
+        }
+
+        $this->setNull('tokens');
+
         $re = '/.*{{(.*)}}/m';
 
         $matches = [];
 
         $tmp = [];
-        preg_match_all($re, $this->get('subject'), $tmp, PREG_SET_ORDER, 0);
+        preg_match_all($re, $this->get('subject') ?? '', $tmp, \PREG_SET_ORDER, 0);
         $matches = array_merge($matches, $tmp);
 
         $tmp = [];
-        preg_match_all($re, $this->get('html'), $tmp, PREG_SET_ORDER, 0);
-        $matches = array_merge($matches, $tmp);
+        preg_match_all($re, $this->get('html') ?? '', $tmp, \PREG_SET_ORDER, 0);
+        $matches = [...$matches, ...$tmp];
 
         $tmp = [];
-        preg_match_all($re, $this->get('text'), $tmp, PREG_SET_ORDER, 0);
-        $matches = array_merge($matches, $tmp);
+        preg_match_all($re, $this->get('text') ?? '', $tmp, \PREG_SET_ORDER, 0);
+        $matches = [...$matches, ...$tmp];
 
-        $tokens = $this->ref('tokens')->export(null, 'token');
         $new_tokens = [];
-
-        //$this->set('tokens', []);
-
-        // @todo can be done better?
-        foreach ($matches as [$match, $token]) {
-            if (in_array($token, $tokens, true)) {
-                continue;
-            }
-
+        foreach ($matches as $match) {
+            $token = $match[1];
             $new_tokens[$token] = [
                 'token' => $token,
                 'description' => $tokens[$token]['description'] ?? '',
             ];
         }
 
-        $this->set('tokens', array_values($new_tokens));
+        $this->ref('tokens')->import(array_values($new_tokens));
     }
 }
