@@ -40,7 +40,7 @@ abstract class GenericTestCase extends BaseTestCase
                     'email_id' => 1,
                     'code' => '',
                     'message' => '',
-                    'timestamp' => (new \Datetime())->format('Y-m-d H:i:s'),
+                    'timestamp' => (new \DateTime())->format('Y-m-d H:i:s'),
                 ],
             ],
             $this->getReflectionModelTableName(MailTemplate::class) => [
@@ -54,7 +54,7 @@ abstract class GenericTestCase extends BaseTestCase
                     'cc' => '[]',
                     'bcc' => '[]',
                     'subject' => 'subject mail for {{token}}',
-                    'text' => str_replace('|', PHP_EOL, $template_body),
+                    'text' => str_replace('|', \PHP_EOL, $template_body),
                     'html' => str_replace('|', '<br/>', $template_body),
                     'attachments' => '[]',
                     'tokens' => '[]',
@@ -69,7 +69,7 @@ abstract class GenericTestCase extends BaseTestCase
                     'cc' => '[]',
                     'bcc' => '[]',
                     'subject' => 'subject mail for {{user.name}}',
-                    'text' => str_replace('|', PHP_EOL, $template_user_body),
+                    'text' => str_replace('|', \PHP_EOL, $template_user_body),
                     'html' => str_replace('|', '<br/>', $template_user_body),
                     'attachments' => '[]',
                     'tokens' => '[]',
